WIP: flip healthz when node is unschedulable
WORK IN PROGRESS:

We probably want this to be "readyz".  Current logic is
behaving as liveness (is it making progress?) which could cause kubelet
to restart it (though we don't seem to install a liveness probe - bad!).

Ideally we have a liveness probe URL which gets updates as kube-proxy
makes progress and a readiness probe URL which should be used for LB
health checks (including externalTrafficPolicy checks).

Pushing filtering to cloud providers may want to be done in a more
code-breaking way to force external providers to consider it (e.g.
changing the interface).  This is a good startibg point until we are
ready to have that chat with sig-cloud-provider (they want this work
anyway).  We can push this as a separate PR without any of the
kube-proxy changes.

Unclear how this intersects with externalTrafficPolicy

Unclear how or if users should be able to configure a drain time